requirements.txt
falcon==1.1.0
gunicorn==19.6.0
Jinja2==2.8.1
line-bot-sdk==1.1.0
git+https://github.com/raven38/deel
